// src/enum/event.enum.ts
export enum InquiryEventEnum {
    CREATED = 'inquiry.created',
    UPDATED = 'inquiry.updated'
}

export enum OrderEventEnum {
    ORDERNEW = 'order.new',
    ORDERPAID = 'order.paid',
    STATUSCHANGE = 'order.statusChange',
    ORDERASSIGN = 'order.assign',
    ORDETRACK = 'order.track'
}
export enum UserEventEnum {
    WALLETADD = 'wallet.add'
}